# public/openapi.yaml
openapi: 3.0.0
info:
  title: MarketTracker
  version: 1.0.0
  description: ''
servers:
  - url: http://localhost:3000
  - url: ''
paths:
...
  /api/symbol:
    post:
      summary: Add Symbol
      responses:
        '201':
          description: Add Symbol
          headers:
            vary:
              deprecated: false
              schema: {}
              example: >-
                rsc, next-router-state-tree, next-router-prefetch,
                next-router-segment-prefetch
            Date:
              deprecated: false
              schema: {}
              example: Thu, 15 Jan 2026 08:58:30 GMT
            Connection:
              deprecated: false
              schema: {}
              example: keep-alive
            Keep-Alive:
              deprecated: false
              schema: {}
              example: timeout=5
            Transfer-Encoding:
              deprecated: false
              schema: {}
              example: chunked
      tags:
        - symbol
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                symbol:
                  type: string
            example:
              symbol: ''
...